Mainクラスの名前変更とコメント挿入
